repopratice/self/01_agent/pyproject.toml
[project]
name = "01-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "openai-agents",
  "openai",
  "python-dotenv"
]